+++ packages/core/src/types/wallet.ts
@@ -64,13 +64,13 @@ export interface Metadata {
 
 export interface AppUrl {
   native?:
-    | string
-    | {
-        android?: string;
-        ios?: string;
-        macos?: string;
-        windows?: string;
-      };
+  | string
+  | {
+    android?: string;
+    ios?: string;
+    macos?: string;
+    windows?: string;
+  };
   universal?: string;
 }
 
@@ -82,14 +82,15 @@ export interface Wallet {
   mobileDisabled: boolean | (() => boolean);
   description?: string;
   rejectMessage?:
-    | {
-        source: string; // message from wallet app
-        target?: string; // message stored in walletManager, default 'Request Rejected!'
-      }
-    | string; // message from wallet app
+  | {
+    source: string; // message from wallet app
+    target?: string; // message stored in walletManager, default 'Request Rejected!'
+  }
+  | string; // message from wallet app
   rejectCode?: number; // code from wallet app
   connectEventNamesOnWindow?: string[];
   connectEventNamesOnClient?: string[];
+  supportedChains?: string[]; // array of supported chains
   downloads?: DownloadInfo[];
   logo?: string | { major: string; minor: string };
   walletconnect?: {
@@ -244,7 +245,7 @@ export interface WalletClient {
 export type WalletAdapter = ChainWalletBase | MainWalletBase;
 
 export interface IChainWallet {
-  new (walletInfo: Wallet, chainInfo: ChainRecord): ChainWalletBase;
+  new(walletInfo: Wallet, chainInfo: ChainRecord): ChainWalletBase;
 }
 
 export type NameServiceName = string;