+++ wallets/leap-capsule-social-login/src/client.ts
@@ -1,16 +1,26 @@
 /* eslint @typescript-eslint/no-explicit-any: 0 */ // --> OFF
 
-import { StdSignDoc } from "@cosmjs/amino";
-import { SignOptions, SignType } from "@cosmos-kit/core";
-import { DirectSignDoc, WalletClient } from "@cosmos-kit/core";
-import { CapsuleProvider } from "@leapwallet/cosmos-social-login-capsule-provider";
-import { SignDoc } from "cosmjs-types/cosmos/tx/v1beta1/tx";
+import {
+  encodeEd25519Pubkey,
+  encodeSecp256k1Pubkey,
+  pubkeyType,
+  StdSignature,
+  StdSignDoc,
+} from '@cosmjs/amino';
+import {
+  DirectSignDoc,
+  SignOptions,
+  SignType,
+  WalletClient,
+} from '@cosmos-kit/core';
+import { CapsuleProvider } from '@leapwallet/cosmos-social-login-capsule-provider';
+import { SignDoc } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 
 export class CosmosCapsuleClient implements WalletClient {
   readonly loginProvider: CapsuleProvider;
 
   constructor(options: { loginProvider: CapsuleProvider }) {
-    this.loginProvider = options.loginProvider
+    this.loginProvider = options.loginProvider;
   }
 
   async disconnect() {
@@ -20,7 +30,7 @@ export class CosmosCapsuleClient implements WalletClient {
   async getSimpleAccount(chainId: string) {
     const { address, username } = await this.getAccount(chainId);
     return {
-      namespace: "cosmos",
+      namespace: 'cosmos',
       chainId,
       address,
       username,
@@ -29,7 +39,7 @@ export class CosmosCapsuleClient implements WalletClient {
 
   async enable() {
     await this.handleConnect();
-}
+  }
 
   async handleConnect() {
     await this.loginProvider.handleConnect();
@@ -40,7 +50,7 @@ export class CosmosCapsuleClient implements WalletClient {
   }
 
   getOfflineSigner(chainId: string, preferredSignType?: SignType) {
-    return this.loginProvider.getOfflineSigner(chainId, preferredSignType)
+    return this.loginProvider.getOfflineSigner(chainId, preferredSignType);
   }
 
   getOfflineSignerAmino(chainId: string): any {
@@ -55,12 +65,58 @@ export class CosmosCapsuleClient implements WalletClient {
     chainId: string,
     signer: string,
     signDoc: StdSignDoc,
-    signOptions?: SignOptions,
+    signOptions?: SignOptions
   ) {
-    return this.loginProvider.signAmino(chainId, signer, signDoc, signOptions)
+    return this.loginProvider.signAmino(chainId, signer, signDoc, signOptions);
   }
 
-  async signDirect(chainId: string, signer: string, signDoc: DirectSignDoc): Promise<any> {
-    return this.loginProvider.signDirect(chainId, signer, signDoc as unknown as SignDoc);
+  async signArbitrary(
+    chainId: string,
+    signer: string,
+    data: string | Uint8Array
+  ): Promise<StdSignature> {
+    const account = await this.getAccount(chainId);
+    if (!account) {
+      throw new Error(`Wallet not connected to chain ${chainId}`);
+    }
+    const pubkey = (() => {
+      switch (account.algo) {
+        case 'secp256k1':
+          return encodeSecp256k1Pubkey(account.pubkey);
+        case 'ed25519':
+          return encodeEd25519Pubkey(account.pubkey);
+        default:
+          throw new Error('sr25519 public key algorithm is not supported');
+      }
+    })();
+
+    const signature = await this.loginProvider.signArbitrary(
+      chainId,
+      signer,
+      data
+    );
+
+    return {
+      signature,
+      pub_key: {
+        type:
+          account.algo === 'secp256k1'
+            ? pubkeyType.secp256k1
+            : pubkeyType.ed25519,
+        value: pubkey.value,
+      },
+    };
+  }
+
+  async signDirect(
+    chainId: string,
+    signer: string,
+    signDoc: DirectSignDoc
+  ): Promise<any> {
+    return this.loginProvider.signDirect(
+      chainId,
+      signer,
+      signDoc as unknown as SignDoc
+    );
   }
 }