+++ wallets/leap-capsule-social-login/src/leap-social.ts
@@ -1,5 +1,6 @@
-import { LeapCapsuleInfo } from "@leapwallet/cosmos-social-login-capsule-provider";
-import { LeapCapsuleWallet } from "./main-wallet";
+import { LeapCapsuleInfo } from '@leapwallet/cosmos-social-login-capsule-provider';
+
+import { LeapCapsuleWallet } from './main-wallet';
 
 const leapSocial = new LeapCapsuleWallet(LeapCapsuleInfo);