+++ wallets/okto-extension/src/okto.ts
@@ -1,0 +1,5 @@
+import { oktoExtensionInfo, OktoExtensionWallet } from './extension';
+
+const oktoExtension = new OktoExtensionWallet(oktoExtensionInfo);
+
+export const wallets = [oktoExtension];