+++ wallets/prax-extension/src/index.ts
@@ -1,0 +1,2 @@
+export * from './extension/index.js';
+export * from './prax.js';