+++ wallets/cdcwallet-extension/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.15.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.15.0...@cosmos-kit/cdcwallet-extension@2.15.1) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
+# [2.15.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.14.0...@cosmos-kit/cdcwallet-extension@2.15.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
+# [2.14.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.13.2...@cosmos-kit/cdcwallet-extension@2.14.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
+## [2.13.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.13.1...@cosmos-kit/cdcwallet-extension@2.13.2) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
+## [2.13.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.13.0...@cosmos-kit/cdcwallet-extension@2.13.1) (2024-06-26)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
+# [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cdcwallet-extension@2.12.0...@cosmos-kit/cdcwallet-extension@2.13.0) (2024-06-18)
+
+**Note:** Version bump only for package @cosmos-kit/cdcwallet-extension
+
+
+
+
+
 # 2.12.0 (2024-06-05)