+++ wallets/cosmos-extension-metamask/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.12.0...@cosmos-kit/cosmos-extension-metamask@0.12.1) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+# [0.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.11.0...@cosmos-kit/cosmos-extension-metamask@0.12.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+# [0.11.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.10.0...@cosmos-kit/cosmos-extension-metamask@0.11.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+# [0.10.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.9.3...@cosmos-kit/cosmos-extension-metamask@0.10.0) (2024-08-01)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.9.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.9.2...@cosmos-kit/cosmos-extension-metamask@0.9.3) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.9.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.9.1...@cosmos-kit/cosmos-extension-metamask@0.9.2) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+## [0.9.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.9.0...@cosmos-kit/cosmos-extension-metamask@0.9.1) (2024-06-26)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
+# [0.9.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.8.0...@cosmos-kit/cosmos-extension-metamask@0.9.0) (2024-06-18)
+
+**Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask
+
+
+
+
+
 # [0.8.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmos-extension-metamask@0.7.0...@cosmos-kit/cosmos-extension-metamask@0.8.0) (2024-05-28)
 
 **Note:** Version bump only for package @cosmos-kit/cosmos-extension-metamask