+++ wallets/keplr-extension/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.13.0...@cosmos-kit/keplr-extension@2.14.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+# [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.12.2...@cosmos-kit/keplr-extension@2.13.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+## [2.12.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.12.1...@cosmos-kit/keplr-extension@2.12.2) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+## [2.12.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.12.0...@cosmos-kit/keplr-extension@2.12.1) (2024-06-26)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
+# [2.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.11.1...@cosmos-kit/keplr-extension@2.12.0) (2024-06-18)
+
+**Note:** Version bump only for package @cosmos-kit/keplr-extension
+
+
+
+
+
 ## [2.11.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/keplr-extension@2.11.0...@cosmos-kit/keplr-extension@2.11.1) (2024-05-29)
 
 **Note:** Version bump only for package @cosmos-kit/keplr-extension