+++ wallets/web3auth/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.13.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.12.0...@cosmos-kit/web3auth@2.13.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+# [2.12.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.11.0...@cosmos-kit/web3auth@2.12.0) (2024-10-09)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+# [2.11.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.10.2...@cosmos-kit/web3auth@2.11.0) (2024-07-11)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+## [2.10.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.10.1...@cosmos-kit/web3auth@2.10.2) (2024-06-28)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+## [2.10.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.10.0...@cosmos-kit/web3auth@2.10.1) (2024-06-26)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
+# [2.10.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.9.0...@cosmos-kit/web3auth@2.10.0) (2024-06-18)
+
+**Note:** Version bump only for package @cosmos-kit/web3auth
+
+
+
+
+
 # [2.9.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/web3auth@2.8.0...@cosmos-kit/web3auth@2.9.0) (2024-05-28)
 
 **Note:** Version bump only for package @cosmos-kit/web3auth